07_django/INSTA/requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.6
bleach==3.1.0
cffi==1.12.2
colorama==0.4.1
cryptography==2.6.1
decorator==4.4.0
defusedxml==0.5.0
Django==2.1.7
django-appconf==1.0.3
django-bootstrap4==0.0.8
django-extensions==2.1.6
django-imagekit==4.0.2
django-rest-framework==0.1.0
djangorestframework==3.9.2
entrypoints==0.3
fabric==2.4.0
Faker==1.0.4
invoke==1.2.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.8
pandocfilters==1.4.2
paramiko==2.4.2
parso==0.4.0
pickleshare==0.7.5
pilkit==2.0
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
PyNaCl==1.3.0
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
pyzmq==18.0.1
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
tornado==6.0.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
